+++ src/Models/Automation.php
@@ -2,11 +2,14 @@
 
 namespace Escalated\Laravel\Models;
 
+use Escalated\Laravel\Concerns\UsesEscalatedConnection;
 use Escalated\Laravel\Escalated;
 use Illuminate\Database\Eloquent\Model;
 
 class Automation extends Model
 {
+    use UsesEscalatedConnection;
+
     protected $guarded = ['id'];
 
     public function getTable(): string